feat: song index page + candidate browser breadcrumb (sections 2-4)
web/app/page.tsx — song index (replaces old candidate browser):
- Fetches GET /songs on mount; redirects to /candidates on 503 (single-song mode)
- Grid of song cards: title, thread slug, key, BPM, color dot, singer
- Card click: POST /songs/activate then push to /candidates; spinner on clicked card
- "No songs found" empty state; error state for unreachable API

web/app/candidates/page.tsx — candidate browser (moved from app/page.tsx):
- All existing functionality preserved unchanged
- Adds breadcrumb: "← Songs / <title>" above heading when GET /songs/active
  returns an active song; hidden in single-song mode (active: null)

web/lib/api.ts — fetchSongs(), activateSong(), fetchActiveSong()
web/lib/types.ts — SongEntry interface

TypeScript build passes; routes / and /candidates both present.
